requirements.txt
matrix-nio>=0.6
Markdown>=3.1.1
PyYAML>=5.1.2
